Linux: Update user AF_RPMSG define for 3.9 kernel
IPC's Linux user space support contains a copy of a few
kernel-defined structs and definitions.  This is necessary
because the kernel headers are not include-able by user space.
One of these values is the socket address family used for
rpmsg-based sockets - AF_RPMSG.  This definition is always
'one more' than the last address family defined by the kernel's
socket.h.

Prior to Linux 3.9, this was aways 40.  In Linux 3.9, a new
address family was introduced (AF_VSOCK), and as a result we
needed to set AF_RPMSG to 41.  To enable this, the user space
header reflects on the kernel version and appropriately assigns
the value of AF_RPMSG.

Note that in addition to adding 3.9+ support, this commit also
cleans up some dead code.