Linux: Added GateMP support for DRA7XX devices
GateMP support has been added using the device's spinlocks.

To enable GateMP on the host, throw -g option when launching the
LAD daemon as follows:

        ./lad_dra7xx log.txt -g

To enable GateMP on the slave core, the core must be configured as
the owner of SR0 and GateMP.hostSupport config param must be set
to 'true' as follows:

         var GateMP = xdc.useModule('ti.sdo.ipc.GateMP');
         GateMP.hostSupport = true;

The slave core with GateMP enabled must the loaded (remoteproc)
prior to the execution of LAD on the host. This adds a restriction
on LAD that wasn't previously there.

Note: Initial validation has only been performed on DRA7XX DSP1.