OMAP5: Remove unnecessary rpmsg-related code
The original OMAP5 code import included support for running the
IPU in non-SMP BIOS.  This was untested in the context of IPC 3.x,
and is not supported.  The unsupported/dead stuff was obfuscating
the live stuff, so this commit removes it.

There is no intentional change in behavior, other than slight
performance improvement due to small bits of code removal.