VAYU QNX: Prevent race condition when ARM is slow to handle mailbox interrupts
This patch fixes a race condition where the host side in QNX is slow to clear a
mailbox interrupt. If the slave were to try to send another interrupt before
the ARM has cleared the first one, the code in InterruptDsp_intSend()
will simply drop it. This is bad because on QNX, the code in VirtQueue_ISR
interprets the received arg as the VirtQueue id. Hence if consecutive
messages were sent with different ids, the second message would be missed.

The fix is to poll whenever the mailbox is occupied, and send the message
only when the mailbox is free. (SDOCM00106278)

Signed-off-by: VW <vwan@ti.com>