DRA7XX: Remove unnecessary rpmsg-related code
Similar to commit 16adcaad which cleaned up OMAP5 code, this
removes dead code for DRA7XX.  Most of the changes remove
untested, and likely broken, non-SMP code.

In at least one case, the code cleanup may fix previously
undiscovered issues.  For example, in VirtQueue_isr(), the first
switch() block was [likely inadvertently] skipped for DSP2.  There
is no known issue because of this, but now with this cleanup, DSP2
will (and should!) participate in this initial switch statement.