BIOS: GateMP host support must ensure maxNameLen is 32
When GateMP host support is enabled, the default GateMP must ensure
the size of the maxNameLen must be at least 32. The size must correspond
to the expected size on the host OS (Linux, Qnx, Android) which is
currently 32.

The default size for maxNameLen on the BIOS-side is 32. Thus this issue
hadn't be encountered.