66AK2E: BIOS: Remove leading 'C' in device name
The initial support for 66AK2E (commit
867efc6fa15f91639820614f6f9ca6ee6751ffa6) went in with
the 'C66AK2E' identifier.  After review by the MCSDK team,
the leading 'C' was incorrect and removed.

The XDC platform and catalog includes this leading 'C', so
most of the internal config-related support did not require
changes.

Finally, this is only the BIOS-side changes.  The Linux-side
changes will be done in a subsequent commit.