Increase NAMESERVER_GET_TIMEOUT from 20000 to 40000 microseconds
Product testing revealed that a remote core was occasionally taking
slighly longer than 20000 microseconds to respond (in the negative)
to a NameServer_getRemote() query.  We're not sure why it's taking
so long in some cases, as normally the remote core will respond in a
few hundred microseconds.

Doubling the timeout will affect the total query time across multiple
remote cores when some of those cores aren't running NameServer, but
this should be OK since it affects "startup" times with MessageQ_open()
and shouldn't affect normal "execute" times.