Test: Fix NameServerApp config
IpcMgr_ipcStartup() was being added twice to the BIOS startup
function - once in nameserverapp.cfg (which is removed in this
commit) and again in messageq_common.cfg.xs (which is included
by nameserverapp.cfg).  In addition to just being wrong, this
resulted in an abort on Keystone 2 devices.

Also, the MessageQ.SetupTransportProxy was errantly not assigned
in the nameserverapp.cfg script.  This resulted in MessageQ.xs
pulling in 'default' MessageQ configuration, which pulled in much
more content than was necessary for this test (for example,
TransportShm, which then pulled in GateMP, SharedRegion, etc).  By
explicitly assigning MessageQ.SetupTransportProxy to the
ti.ipc.transports.TransportRpmsgSetup module, these extra [unused]
modules are no longer pulled in.