Eliminate need to specify Syslink_Override_Params on QNX
This commit changes Syslink_Override_Params in Platform.c to an empty
string. It removes the need to specify carveouts in the string, which
was only used for verification purposes when there is a slave-side MMU.
Code associated with the verification of the resource table entries
against these carveouts is removed.

This commit also changes the default for mmuEnable to TRUE, since that
is currently the only mode we support.

This addresses CQ SDOCM00105667.

Signed-off-by: VW <vwan@ti.com>