Remove dead code that appears to be trying to set L1D cache size to 0
The way to set an L1/L2 cache size is through the Build.platformTable
platform instance creation.  The dead code that is removed here was not
only trying to change the cache size at too late a time, it was doing
so by setting a local variable that went nowhere.