Fix VirtQueue_getAvailBuf on QNX to avoid race condition
This is to address SDOCM00106156.

In commit c98d2c5624aac20bcdd738f5a54fbddca045fd69, we have fixed a race
condition on the BIOS side by clearing the VRING_USED_F_NO_NOTIFY flag
prior to checking whether there are remaining buffers. We are now doing
the same on the host side.

Note that this code is not currently exercised on the host, but is simply
maintained to ensure that the VirtQueue module is complete in functionality.

Signed-off-by: VW <vwan@ti.com>