Create Keystone platform instances with no external DDR defined
On MAR-based cores (i.e., C6x), SYS/BIOS will, by default, set the
"cacheable" bit for any external memory that it "sees", i.e., any
external memory that is defined for the platform.  This happens regardless
of whether or not that memory is used for program placement.  This
cacheable-ness is somewhat hidden to the programmer.

In order to offer better control of this, Keystone platform instances are
created with no external memory defined.  This allows all MAR pages to be
tagged "non-cached" by default, and the programmer is free to set MAR bits
to their liking.