Fix MessageQ to support more than 52 instances in QNX
This fix changes MessageQ to obtain a port number assignment from MessageQCopy
and use it as local queue id as opposed to having MessageQ manage the ids
itself. This ensures the port number to be above 1024, so that it does not
collide with any reserved port such as the one for the Name Service.

This addresses CQ SDOCM00104834.

Signed-off-by: VW <vwan@ti.com>