Linux/Android: Option to set LAD working dir permissions
Added an optional flag (-p) to LAD, to allow users to set the permissions
of LAD's working directory at launch.

This is critical, if applications using IPC stack (MessageQ) are executed
in user-space while LAD is started as root.  Without proper directory
permissions the user application will be unable to talk to the LAD daemon.