Add INTMEM resource type support in QNX
This commit adds support for the new INTMEM resource type (SDOCM00107479).

For this resource type, we map the entry only for the loader on the host,
without mapping it to the slave MMU.

This commit also removes support of the hard-coded static entries defined
in the address translation table, which was originally there to ensure
slave internal memories are always mapped. To match up with the Linux
implementation we'll now rely on the INTMEM resource type instead.

Note that this does break compatibility with slave executables that were
previously built without the INTMEM resource entries but were relying
on the hard-coded internal memory definitions in the ipc resource manager.

Signed-off-by: VW <vwan@ti.com>