SDOCM00108114 Link error when building with DCE on IPU/Vayu
The notify setup module for vayu was unconditionally pulling in the
notify shared memory driver which, in turn, pulled in the ti.sdo.ipc
package. This caused undefined references when building an executable
for DCE because it was not supposed to use any IPC modules. The notify
driver now has a check to see if it should pull in IPC modules or not.