Set Hwi.excHandlerFunc earlier, before it's sealed
One of the newly-used SYS/BIOS or XDC releases introduced a change
that caused the setting of Hwi.excHandlerFunc in module$static$init()
to fail, due to it being "sealed" at that time.  Moving it to an
earlier-invoked module method solves this issue, albeit with a caveat
that is mentioned above where the Hwi delegate is retrieved.

This issue was seen when moving from
	xdctools_3_25_05_94 -> xdctools_3_30_01_25
	bios_6_37_02_27 -> bios_6_40_01_15