Support in QNX MmRpc/MmServiceMgr to identify deleted instance during cleanup
This commit adds the MmRpc_getId API on host to return a unique id for
each MmRpc instance of a service. This id is passed to the cleanup function
registered in the newly introduced MmServiceMgr_register2() so that it can
clean up after a given MmRpc instance.

This commit addresses SDOCM00108138

Signed-off-by: VW <vwan@ti.com>