Add option to provide a valid path to crash dump file in QNX
This commit fixes an error that occurs during the recovery process when
an invalid path is used to store the contents of the remote core logs.
The -c flag can now be passed to the resource manager to specify a valid
path for the file.

This also changes a few instances where error messages were referring to
"syslink".

It addresses SDOCM00108056.

Signed-off-by: VW <vwan@ti.com>