Move Watchdog module configuration assignments to module$use()
When the ti.deh package is loaded the Watchdog module$meta$init() method
is run.  If the package is loaded by an architecture/device that is not
supported by Watchdog then module$meta$init() fails and the whole
configuration fails.  This happens even when the Watchdog module is *not*
used, such as when the ti.deh package gets loaded as a result of an
xdc.module() reference.