SDOCM00113338 fix: reduce number of Tasks for messageq_multi.c
Also reduce the maximum number of threads supported in MessageQMulti.c

The need for this fix arose when the BIOS heap size was reduced in
some .cfg files in packages/ti/ipc/tests.  The new, smaller heap was
not sufficient for allocating 55 Tasks (it failed at about the 30th
Task).