SDOCM00113341 fix: restrict writing Ipc_Reserved struct to core's own area
In Ipc_attach(), when a 'memReq' size is 0, a core can write to another
core's Ipc_Reserved struct in SR0 when it isn't supposed to *ever* write
another core's area.  The incorrect assignment needs to be done only when
MultiProc_self() < remoteProcId, similarly to how it is done in other places,
which ensures that a core is writing to only its own IpcReserved area.