Linux/MessageQ: Close the socket upon failure
The socket created in transportCreateEndpoint is not freed up
immediately if the SocketBindAddr fails, and is not freed up
by the application in the corresponding MessageQ_delete as well.
These unbound sockets only get freed up as part of the process
cleanup of any open file descriptors by the Linux kernel. Close
these sockets proactively, so that there are no stale sockets
present during the execution of a process.

Signed-off-by: Suman Anna <s-anna@ti.com>