SDOCM00113753 Notify mailbox driver is clearing wrong fifo
Fixed the notify mailbox driver to clear the inbound fifo, not the
outbound fifo. In addition, write the End-Of-Interrupt (EOI) register
after clearing the fifo; this was not being done. Finally, after
fetching inbound messages (in the notify ISR), the EOI register was
not written correctly; now it writes the corresponding interrupt index.