Move cluster processor list from config to module state
The list of processors in the cluster used to be defined at
configuration time. This prevents reusing a single executable
image on multiple processors (i.e. one image for all DSPs).
I moved this list into the MultiProc module state, so that it
could be defined in a reset hook when reusing a single image.