Rename ipc-trace paths to /dev/ipc-trace/<core name> in QNX
Renaming the trace paths to match the convention adopted by /dev/ipc-file
and /dev/ipc-state, so that MultiProc id is not required to be known ahead
of time.

Note that this introduces a break in compatibility for anyone who's been
relying on the path name /dev/ipc-trace#.

This addresses SDOCM00113638.

Signed-off-by: VW <vwan@ti.com>