Fix GateMP to allow SR0's physical address to be different from its virtual
This commit fixes GateMP host-support (when host is running an HLOS) to allow
SR0's physical address to be a different value from its virtual address on the
slave core, by performing a virtual to physical address translation on the
addresses of the inUse arrays prior to stashing them into a NameServer entry
on the slave.

This addresses CQ SDOCM00113149.

Signed-off-by: VW <vwan@ti.com>