Remove references to SysLink in QNX code base
This commit removes all references to "SysLink" in the QNX IPC code base, with
the exception that it preserves the file and directory names. The decision was
made to preserve the path names to avoid breaking any existing apps that may
be including headers and/or libraries based on the existing directory
structure, and to ease git history lookup going forward.

Note that this does introduce a break in the naming of internal features such
as SYSLINK_DEBUG and SYSLINK_DEBUG_TRACE_LEVEL.

This helps address SDOCM00108142.

Signed-off-by: VW <vwan@ti.com>