Simplify configuration of maximum name length in NameServer
This commit cleans up how the maximum name length is configured in NameServer.
Now the user only needs to configure it during instance creation, and to
configure it at the remote level in
packages/ti/ipc/namesrv/_NameServerRemoteRpmsg.h.

When a name passed to the NameServer remote exceeds the length it can support,
it now returns NameServer_E_NAMETOOLONG.

The NameServerApp is updated to test for long names that exceed the
maximum length.

This commit addresses SDOCM00111686.

Signed-off-by: VW <vwan@ti.com>