Specify Address Family for RPMessage in products.mak
We use the Linux kernel version to make a guess at the value
of AF_RPMSG. However, depending on the kernel configuration,
this guess could be wrong. In this case, you must specify the
AF_RPMSG value directly. This had been done in ipc-linux.mak,
but this was too obscure. So, I moved it to products.mak.