Remove MAX #define from Linux Std.h
The MAX() macro that is #defined in Std.h for Linux can either conflict
or be a duplicate of other MAX() macro definitions.  GCC's glibc defines
MAX in sys/param.h, and code builds will either warn abour or error on
a duplicate or conflicting definition.

Remove the usage of MAX completely to avoid this.  Use a locally-defined,
differently-named macro in C file if appropriate.