SDOCM00115434 NameServer_delete in LAD daemon leaks memory
Remove invalid assertion that name list must be empty. This
is not true and instance might still be in use by another
client (object is reference counted). Add while loop to delete
each entry on the name list. Destroy the object's gate (was not
being done). Remove macro to destruct a list object (not needed).