Add capability to specify specific Watchdog timer for Deh
Previous to this change the user was basically stuck using the GPTimer
that was hardcoded to their device (although they could change the
timer by changing the device's timer settings, which was cumbersome).

The user can now specify a particular GPTimer by setting it in the .cfg
file.  For DRA7XX's IPU2, the Watchdog timers (one for each SMP core)
can be set with:
    Watchdog.timerIds.length = 2;
    Watchdog.timerIds[0] = "GPTimer9";
    Watchdog.timerIds[1] = "GPTimer4";

Each supported device has default GPTimer settings appropriate to that
device (the snippet above actually assigns the current defaults for
Vayu, so it is actually a no-op).

The "GPTimer<n>" specification encapsulates timer settings such as
intNum and eventId.  Many GPTimer specifications in Watchdog.xs have
-1 in these fields if they don't have a hardcoded wirings for them, yet
they can still be used if the user explicitly defines the value in the
.cfg file and somehow sets up the hardware to generate that intNum or
eventId (e.g., through the crossbar on Vayu).