Add GateHWSpinlock configuration into qnx/src/cfg
This commit adds a configuration structure for GateHWSpinlock into the
qnx/src/cfg directory. Some refactoring of the GateHWSpinlock implementation
is necessary to allow both the user library and the resource manager to
share this common configuration. ipc3x_dev/ti/syslink/Std.h is cleaned up to
eliminate duplication with respect to ti/ipc/Std.h and to fix resulting
build errors.

Signed-off-by: VW <vwan@ti.com>