Add new command-line options to LAD to override static defaults
Add -n <nprocs> support for overriding the total number of processors
in the system (not to be confused with number of processors in cluster).

Add -r <nreserve> support for overriding the number of reserved
MessageQ queues.

Add -s <synctype> support for overriding the Ipc_ProcSync_* type (can
be -s ALL|PAIR|NONE).