Change default polling interval for default gate in QNX to 10 ms
This commit decreases the default polling interval for the default gate during
GateMP_setup to 10 ms. This is so that slave executables that may take a bit
longer to boot up and create the default gate do not cause the IPC driver to
wait for a full second before retrying the poll, reducing the driver
startup time in such cases.

Signed-off-by: VW <vwan@ti.com>