Add power management support to DRA7xx DSP (part 2)
This commit ties up a few loose ends that were introduced in a previous commit
to add PM support for the DSP:

- Add a workaround for a silicon bug in which the DSP would hang when the idle
instruction is invoked while it is placed in external memory and prefetch is
enabled
- Remove unnecessary Swi thread and call IpcPower_suspend in Task context
- Add support for DSP2
- Add a null implementation for IpcPower that is compiled in when BIOS Power
module is not used. This is simply to satisfy all existing call sites that
are calling into IpcPower regardless if PM is enabled. It also prevents the
BIOS Power module from being unnecessarily linked in when it is not needed.
A possible future enhancement would be to look into compiling out these
call sites when PM is not enabled.

This is towards addressing CQ SDOCM00117446.

Signed-off-by: VW <vwan@ti.com>