QNX: Modification for QNX 6.5 compilation
This patch introduces the changes needed so that
the QNX IPC can be compiled against QNX SDP 6.5.0.

Where needed, a check is made to make one call
when compiling against QNX SDP 6.6 and another
call when compiling against QNX SDP 6.5 in order
to take advantage of the newer functionality of
QNX SDP 6.6.

Signed-off-by: Buddy Liong <a0270631@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>