Tests: ping_rpmsg: Update Test with Proper Socket Usage
This test was incorrectly using a socket that was connected
to a remote endpoint to receive messages on the HOST. The
remote core was using the source address to send the response
when communicating with an rpmsg-proto channel, which is not
a valid use of the rpmsg-proto channel. Only sockets which
have been bound to a local endpoint by the application should
be used to receive messages.

When the remote core sends a message to an endpoint that
is associated with a channel, there is a kernel crash because
the private data for that endpoint is not as expected.

Update the test case to properly use rpmsg-proto. With these
updates, the test case now passes and there is no kernel
crash.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>