daemon: Add support for UIO-based SR0
The GateMP daemon uses /dev/mem to access SR0, where the gate
usage arrays are kept.  /dev/mem may not be present in all
production systems due to the security risk that it implies.

A userspace I/O driver approach can be used as an alternative.
The Linux kernel needs to expose a UIO named 'sr0' in order
to access the shared region from userspace.  The daemon is now
capable of finding such UIO entry if exists, and otherwise
fall back to the previous /dev/mem based approach.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>