Linux: Add support for the hwspinlock_user driver
The hwspinlock_user is a character driver that exposes ioctls to
lock and unlock a given hwspinlock instance.  The hwspinlock_user
creates the /dev/hwspinlock device file.

The hwspinlock_user driver is used as the first option when available,
otherwise the previous /dev/mem approach is used.

The hwspinlock_user.h uapi header file is kept locally as it may not
be present in all supported kernels.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>