lad: Clean-up old response FIFOs
The LAD daemon was only deleting specific FIFOs for new clients
whose PIDs had a corresponding old FIFO left over.  This was not
effective as the client side checks if the FIFO doesn't exist
before sending the LAD_CONNECT command, so the old FIFO was
not deleted.

The cleanup is changed to delete any old FIFO left over from
previous LAD sessions.  The cleanup implementation is based on
wpa_supplicant's.

Signed-off-by: Vishal Mahaveer <vishalm@ti.com>
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>