tests: Update messageq_multicore to use ti.ipc.remoteproc.Resource
Before this fix, the rsc table header file was included directly in C
code. And the table was getting optimized away due to no direct
reference to the resource table. This resulted in failure of test
downloaded through user space loader.

Instead added the configuration to include the resource table (without
vdev and vring) through configuration file, which is more cleaner.

Signed-off-by: Sam Nelson <sam.nelson@ti.com>