deh: Watchdog: Timer structures declared as volatile
Watchdog implementation configures timers to operate in posted-write
mode, requiring SW to poll on TWPS register prior to subsequent write
access.

Therefore, references to the Watchdog timer instance must be declared
as volatile.  Otherwise, the initial value read of the TWPS will be
used as the polling value indefinitely, causing a potential lockup
if the application is completely idle.

Signed-off-by: Stephen Molfetta <sjmolfetta@ti.com>