Linux: GateMP: Baseaddr not Correctly Set when UIO is not Present
In the case where UIO is not present, the GateMP_get_sr0() function
should set the baseaddr to 0. However, the code was not correctly
de-referencing the baseaddr parameter when attempting to set it to
zero, resulting in an un-initialized value for baseaddr when the UIO
driver is not present. This can result in an incorrect address for
the inUse arrays with consequences such as failure to map or other
unexpected behavior.

This patch fixes it to properly set the baseaddr to 0 for the case
where UIO is not present.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>