Remove INTMEM Type to match Linux K4.x
Starting in Linux kernel 4.1, the INTMEM type is removed
in favor of the internal memory sections needed for loading
being handled entirely in the Linux kernel. As such, the
INTMEM type no longer is defined.

This patch updated the rsc_types header file to match the
Linux defined types, which no longer defines the INTMEM
type.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>