grcm: Only Assert on Received Message Checks if Shutdown is False
If RPMessage_recv returns due to shutdown, then no valid message
is received by the call, and it should not assert if the message
is invalid.

This patch fixes the code to only do the assert checks if shutdown
is false.

Also fixes a minor issue with a trace referencing a parameter
that no longer exists.

Signed-off-by: Angela Stegmaier <angelabaker@ti.com>